Rename CPU reductions; separate reduction functors from kernels. * Reductions for CPU renamed with CPU suffix. * Created a new file, reductions.h, which contains pariwise reductions.
